move Flush to the test_visibility module
